Move list into primitives
