+++ library/Rattletrap/PropertyValue.hs
@@ -3,7 +3,7 @@ module Rattletrap.PropertyValue where
 import Rattletrap.Primitive.Dictionary
 import Rattletrap.Primitive.Float32
 import Rattletrap.Primitive.Int32
-import Rattletrap.List
+import Rattletrap.Primitive.List
 import Rattletrap.Text
 import Rattletrap.Word64
 import Rattletrap.Word8
